add lesson 22 hw

+++ src/lesson13/Demo.java
@@ -8,7 +8,7 @@ public class Demo {
 
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
 
-        User user = new User(1001, "Ann", "1w21212");
+        User user = new User((long) 1001, "Ann", "1w21212");
         userRepository.save(user);
 
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
@@ -24,7 +24,7 @@ public class Demo {
 
         int n = 15;
         while (n > 0) {
-            User user1 = new User(n, "Ann", "1w21212");
+            User user1 = new User((long) n, "Ann", "1w21212");
             System.out.println(userRepository.save(user1));
             n--;
         }
@@ -34,7 +34,7 @@ public class Demo {
         userRepository.save(null);
 
         //test update
-        user = new User(1001, "Ann", "eretertert");
+        user = new User((long) 1001, "Ann", "eretertert");
         userRepository.update(user);
 
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
@@ -43,7 +43,7 @@ public class Demo {
         //когда нет юзера на обновление
         //когда обновляем null
 
-        user = new User(9999, "Ann", "eretertert");
+        user = new User((long) 9999, "Ann", "eretertert");
         System.out.println(userRepository.update(user));
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
