+++ src/lesson13/Test.java
@@ -148,7 +148,7 @@ public class Test {
 
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
 
-        User user = new User(1001, "Ann", "1w21212");
+        User user = new User((long) 1001, "Ann", "1w21212");
         userRepository.save(user);
 
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
@@ -163,7 +163,7 @@ public class Test {
 
         int n = 15;
         while (n > 0) {
-            User user1 = new User(n, "Ann", "1w21212");
+            User user1 = new User((long) n, "Ann", "1w21212");
             System.out.println(userRepository.save(user1));
             n--;
         }
@@ -173,7 +173,7 @@ public class Test {
         userRepository.save(null);
 
         //test update
-        user = new User(1001, "Ann", "eretertert");
+        user = new User((long) 1001, "Ann", "eretertert");
         userRepository.update(user);
 
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
@@ -182,7 +182,7 @@ public class Test {
         //когда нет юзера на обновление
         //когда обновляем null
 
-        user = new User(9999, "Ann", "eretertert");
+        user = new User((long) 9999, "Ann", "eretertert");
         System.out.println(userRepository.update(user));
         System.out.println(Arrays.deepToString(userRepository.getUsers()));
